Backend/requirements.txt
#asgiref==3.4.0
Django==3.2.4
#pytz==2021.1
sqlparse==0.4.1
djangorestframework==3.11.2
pyjwt==2.1.0
